requirements.txt
nltk==3.8.1
pydub==0.25.1
selenium==4.12.0
undetected-chromedriver==3.5.3
python-dotenv==1.0.0
praw==7.7.1
moviepy==1.0.3
Pillow==10.0.0
pvleopard==1.2.2
instagrapi==1.19.8
ffmpeg